kv260-dp: Set package arch to k26
Override the default package arch inherited from kv260-firmware.inc which
is k26-kv. In order to add this package into the starter rootfs, the arch
is changed to k26 instead. Do this only for this package as kv260-dp will
be used as the default firmware to be loaded by dfx-mgrd during init.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>