som-dashboard: Update init script name
Update the init script name to match the systemd unit file. This will avoid
having the systemd-sysv-generator create a systemd unit file for the init script
on the fly which in turn can lead to two instances of som-dashboard being
started. Also update the SRCREV as the init script resides in the source repo.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>