Add support for K26 Starter Kit SD wic image
This code defines two new machines, k26-som and kv260-starter-kit, the
latter of the two inherits the former and is intended to build a
bootable SD wic image.

All of the included .dts[i] files have been extracted from the latest
Petalinux build to enable the creation of this bootable image but these
files will need to be generated in the future. There are also a number
of small bbappends for device-tree, external-hdf and u-boot-zynq-src as
well as a new .wks file.

Signed-off-by: John Toomey <jtoomey@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>