Create K26 SOM specific bootbin
The Kria SOM bootbin file requires a different type of device tree file
and an additional version header as well as a platform specific BIF file
to include these changes.

The device tree FIT image is generated by u-boot-xlnx but requires some
changes to be made available to the bootbin recipe. An additional recipe
is required to generate the bootbin version header file which is also
included in the generated BIF file.

Signed-off-by: John Toomey <jtoomey@xilinx.com>

Cleanup the k26 u-boot-xlnx entries to be board specific.  This is
primarily done by removing ?= with override specific '=' values.

Signed-off-by: Mark Hatle <mhatle@xilinx.com>