Add k26-starter-kit combined kv/kr image
Add a k26-starter-kit machine config and matching k26-starter-kit-sd.wks
file to support the generation of a combined image capable of booting
both K26 KV and KR board variants. This makes all required dtb files
abailable in /boot and uses a modified boot.scr command to chooses the
correct file by first probing the device ROM.

The use of a uEnv.txt file is not required for any of the Kria images so
it has been removed.

Signed-off-by: John Toomey <jtoomey@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>