Clean up Kria system-{user,conf}.dtsi files
The system-user and system-conf dtsi files were originally imported
directly from the Petalinux BSPs and contain small differences depending
on which one is used. They also contain errors and code which is not
used in the Yocto flow and can be confusing to understand.

This change simplifies these files into a single file which will work on
all platforms and contains only the kernel bootargs (command line) and
stdout-path which is used by u-boot to set the correct UART for output.

This change also resolves an issue where the device-tree used in the
boot.bin does not send console output to the correct UART.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit b26aa59b4d17fc374c195b1c9270aad313cb3bdf)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>