kria-image.bbclass: Create base kria-image bbclass
Move to an image class instead of various petalinux specific image bbappends.

This will allow use to work with custom image recipes, as well as the
existing PetaLinux set.

The IMAGE_CLASSES is always enabled by the kria-common.inc, but will be benign
other then the post rootfs fancontrol which is required for k26.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>