Install correct device tree for KV/KR images
An issue with the way /boot is populated means that the device tree used
in the kv and kr specific images is the system-top.dtb which does not
contain the board specifc device tree entries. This change results in
the correct, board specific device tree being installed over the
system.dtb symlink in /boot

This change is required because the default u-boot script (boot.scr) is
set to boot using /boot/system.dtb so this file must be the correct
device tree for the kv or kr board it is booting on.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>