device-tree: Allow u-boot branch/srcrev to be changed
The device-tree recipe clones the u-boot repo in order to get access to
some of the Kria DTS files but the branch and srcrev user are fixed in
the bbappend. This changes allows these values to be overwritten
elsewhere so that we can use different values for some of the machines.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>