xilinx-bootbin: Version the QSPI bootbin
Create special manifest files for the components of the QSPI.  These are
used to know what went into each of the individual items.  Additionally
the version file is generated based on the 'bootbin-version-header' PV.

Signed-off-by: John Toomey <john.toomey@amd.com>

Move from original XILINX_VER_MAIN to PV, adding the XILINX_VER_BUILD on
non-release/update builds for tracking.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>