kria-qspi: Add manifest file
Add a manifest file to the output of the QSPI recipe containing the
version information of the QSPI as well as the manifests from imgsel and
imgrcry. Also append the full boot.bin manifest as the QSPI contains two
copies of the boot.bin.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>
(cherry picked from commit 452288f4a26addf6694d4c15ad86576a454ede22)
Signed-off-by: Mark Hatle <mark.hatle@amd.com>