u-boot-xlnx: Update the SCR_URI assignment operator
kria specific configuration are getting appended on user configuration
because of ":append" overriding syntax which has higher priority
compared to "+=" operator. Also overriding using ":append" is not
recommended so updating the SRC_URI assignment operator to "+=".

Signed-off-by: Shubhangi Mahalle <shubhangi.shrikrushna-mahalle@amd.com>

Moved from += to .= and excluded the new variable from hash calculations.
This should allow the layer to be removed without impacting recipe hash.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>