axi-qos,ddr-qos: Fix install path
This change is required due to two pull requests in the axi-qos and
ddr-qos repos which is required to fix an install path issue in packages
generated for Ubuntu. This change removes the passing of a specific
install path allowing the recipe to use the default install path which
is set in CMakeLists.txt.

https://github.com/Xilinx/ddr-qos/pull/1
https://github.com/Xilinx/axi-qos/pull/1

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>