generate-machines-sdt.sh: Resolve path to domain.yaml
The script works fine when run from inside the scripts folder. When
the script is called from a different location, the path to the
domain.yaml file does not get resolved properly. Unfortunately,
gen-machine-conf does not warn the user about it if a file is passed
that does not exist, it just silently continues and doesn't use the
domain.yaml file which as a result then does not include any of the
openamp modifications that are defined in the domain.yaml.

Fix this by expanding the path to domain.yaml to an absolute path.

Signed-off-by: Ben Levinsky <ben.levinsky@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>