boot.cmd.kria: Remove unused bootargs
Remove an unused xilinx_tsn_ep.st_pcp bootarg as it is not functional
and could cause confusion.

The name of the kernel module has changed from xilinx_tsn_ep to
xilinx_emac_tsn and the value being passed here matches the default
value in the module so passing it is no longer required. If a user
wishes to change this value they will need to update the bootargs with
the following -

        xilinx_emac_tsn.st_pcp=N

where N is the value they wish to pass.

Signed-off-by: John Toomey <john.toomey@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>