lmsensors-config-kria: Add Kria-specific lmsensors config
Add a Kria-specific lmsensors-config recipe which provides a fancontrol
config file instead of bbappending the defaul lmsensors-config recipe
which provides a dummy fancontrol config file.

The reason for creating a new config recipe instead of having a bbappend
with a machine override is we are building disk images with generic
machines instead of board-specific machines. Hence, we cannot trigger
machine-specific configuration through overrides, instead we create
packages for specific boards that can be installed into board-specific
images.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>
Message-ID: <20251003004442.3289609-5-chris.kohn@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>