openamp-fw-examples: allow for any kria machine
Unlike the versal machines, whose builds are performed entirely using
the EDF flow, the kria builds are still a mix of the old style and EDF.

In a fully EDF build the filesystem can only be built using a generic
machine and the flash component (OSPI/QSPI) is built only using a
specific machine. For kria builds the common filesystem can be performed
with either a common kria machine or a specific one. Therefore change
the override in the OpenAMP firmware example recipe to allow both.

Signed-off-by: Trevor Woerner <trevor.woerner@amd.com>
Message-Id: <20251009143840.756825-1-trevor.woerner@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>