k2{4,6}-smk-sdt.yaml: Enable SYSTEM_DTB_BLOB
In the Kria-QSPI flow, we are packaging the fit-dtb.blob instead of
the plain DTB file. This blob is generated using fdtoverlay by
combining the base DTB and the CCDTBO.
Currently, the default process uses the DTB from the U-Boot build
directory instead of the one generated by the device-tree recipe.
As a result, any updates made to the base using DTSI/YAML files are
not reflected in the fit-dtb.blob.

To ensure the fit-dtb.blob uses the DTB produced by the device-tree
recipe, we need to set the SYSTEM_DTB_BLOB variable.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>
Message-Id: <20251017104430.1431583-1-rajukumar.pothuraju@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>