kria-common.inc: Build board-id-data-* in machine builds
Previously the board-id-data files were built and deployed as part of the
filesystem.  Since the filesystem may be built with a more generic config
we need a way to ensure these files are present during the machine
firmware build.

This implementation may change in the future, but this is a safe way
to ensure that these are always built along side the xilinx-bootbin or
anything that depends on xilinx-bootbin (kria-qspi).

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Message-ID: <20251216005212.2874735-3-mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>